% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nyc_boros_sf}
\alias{nyc_boros_sf}
\alias{nyc_county_sf}
\title{New York City borough boundaries}
\format{
\subsection{\code{nyc_boros_sf}}{

A data frame with 5 rows and 6 columns:
\describe{
\item{boro_code}{Numeric borough code}
\item{boro_name}{Borough Name}
\item{geoid}{Census county GEOID (state + county FIPS, e.g. "36061" for
Manhattan). Character. Use to join county-level Census and ACS tables,
such as those in the nycdemog package.}
\item{shape_leng}{Shape length}
\item{shape_area}{Shape area}
\item{geometry}{Multipolygon}
}
}
}
\source{
\url{https://www.nyc.gov/content/planning/pages/resources#datasets}
}
\usage{
nyc_boros_sf

nyc_county_sf
}
\description{
Clipped to shoreline. EPSG:2263, NAD83 / New York Long Island (ftUS). An sf object.
}
\details{
Borough boundaries for NYC, clipped to shoreline. Produced by NYC Planning Department. Release: 25C, August 2025.

\code{nyc_county_sf} is an identical copy of \code{nyc_boros_sf}, named for
consistency with Census conventions, where the boroughs are counties.
}
\author{
Kieran Healy
}
\keyword{datasets}
