% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nyc_census_tracts_2010_sf}
\alias{nyc_census_tracts_2010_sf}
\title{New York City Census Tracts 2010}
\format{
\subsection{\code{nyc_census_tracts_2010_sf}}{

A simple feature collection with 2165 rows and 12 columns:
\describe{
\item{ct_label}{Census tract identifier}
\item{boro_code}{Borough code as character (1=Manhattan, 2=Bronx, 3=Brooklyn, 4=Queens, 5=Staten Island)}
\item{boro_name}{Borough name}
\item{ct2010}{2010 census tract number (6-digit)}
\item{boro_ct2010}{Unique identifier for 2010 census tracts (BoroCode + CT2010)}
\item{cd_eligibil}{Community District eligibility (E=Eligible, I=Ineligible)}
\item{nta_code}{Unique alphanumeric identifier for each Neighborhood Tabulation Area}
\item{nta_name}{Neighborhood Tabulation Area name}
\item{puma}{Public Use Microdata Area identifier}
\item{shape_leng}{Shape length}
\item{shape_area}{Shape area}
\item{geometry}{Multipolygon}
}
}
}
\source{
\url{https://www.nyc.gov/content/planning/pages/resources#datasets}
}
\usage{
nyc_census_tracts_2010_sf
}
\description{
Simple feature collection of NYC Census Tracts from 2010.
EPSG:2263, NAD83 / New York Long Island (ftUS).
}
\details{
Census tracts are small, relatively permanent statistical subdivisions of a
county. Produced by NYC Planning Department. Release: 25C, August 2025.
}
\author{
Kieran Healy
}
\keyword{datasets}
