% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nyc_congressional_districts_sf}
\alias{nyc_congressional_districts_sf}
\title{New York City Congressional Districts}
\format{
\subsection{\code{nyc_congressional_districts_sf}}{

A simple feature collection with 13 rows and 4 columns:
\describe{
\item{cong_dist}{Congressional district number}
\item{shape_leng}{Shape length}
\item{shape_area}{Shape area}
\item{geometry}{Multipolygon}
}
}
}
\source{
\url{https://www.nyc.gov/content/planning/pages/resources#datasets}
}
\usage{
nyc_congressional_districts_sf
}
\description{
Simple feature collection of NYC Congressional Districts.
EPSG:2263, NAD83 / New York Long Island (ftUS).
}
\details{
Boundaries of NYC Congressional Districts. Produced by NYC Planning Department.
Release: 25C, August 2025.
}
\author{
Kieran Healy
}
\keyword{datasets}
