% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nyc_election_precincts_sf}
\alias{nyc_election_precincts_sf}
\title{New York City Election Precincts}
\format{
\subsection{\code{nyc_election_precincts_sf}}{

A simple feature collection with 4264 rows and 4 columns:
\describe{
\item{elect_dist}{Election district identifier}
\item{shape_leng}{Shape length}
\item{shape_area}{Shape area}
\item{geometry}{Multipolygon}
}
}
}
\source{
\url{https://www.nyc.gov/content/planning/pages/resources#datasets}
}
\usage{
nyc_election_precincts_sf
}
\description{
Simple feature collection of NYC Election Precincts.
EPSG:2263, NAD83 / New York Long Island (ftUS).
}
\details{
Boundaries of NYC Election Districts (also known as Election Precincts).
Produced by NYC Planning Department. Release: 25C, August 2025.
}
\author{
Kieran Healy
}
\keyword{datasets}
