% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nyc_nta10_sf}
\alias{nyc_nta10_sf}
\title{New York City Neighborhood Tabulation Areas 2010}
\format{
\subsection{\code{nyc_nta10_sf}}{

A simple feature collection with 195 rows and 8 columns:
\describe{
\item{boro_code}{Borough code (1=Manhattan, 2=Bronx, 3=Brooklyn, 4=Queens, 5=Staten Island)}
\item{boro_name}{Borough name}
\item{county_fips}{Federal unique ID for counties (005=Bronx, 047=Kings, 061=New York, 081=Queens, 085=Richmond)}
\item{nta_code}{Unique alphanumeric identifier for each Neighborhood Tabulation Area}
\item{nta_name}{Full NTA name}
\item{shape_leng}{Shape length}
\item{shape_area}{Shape area}
\item{geometry}{Multipolygon}
}
}
}
\source{
\url{https://www.nyc.gov/content/planning/pages/resources#datasets}
}
\usage{
nyc_nta10_sf
}
\description{
Simple feature collection of NYC Neighborhood Tabulation Areas from 2010.
EPSG:2263, NAD83 / New York Long Island (ftUS).
}
\details{
Neighborhood Tabulation Areas (NTAs) are aggregations of census tracts that
are subsets of New York City's 55 Public Use Microdata Areas (PUMAs). Produced
by NYC Planning Department. Release: 25C, August 2025.
}
\author{
Kieran Healy
}
\keyword{datasets}
