% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nyc_shoclip_sf}
\alias{nyc_shoclip_sf}
\title{New York City and locale shorelines}
\format{
\subsection{\code{nyc_shoclip_sf}}{

A data frame with 7,426 rows and 16 columns:
\describe{
\item{SOURCE_ID}{DESCRIPTION}
\item{SRC_DATE}{DESCRIPTION}
\item{HOR_ACC}{DESCRIPTION}
\item{INFORM}{DESCRIPTION}
\item{ATTRIBUTE}{DESCRIPTION}
\item{VER_DATE}{DESCRIPTION}
\item{SRC_RESOLU}{DESCRIPTION}
\item{DATA_SOURC}{DESCRIPTION}
\item{EX_METH}{DESCRIPTION}
\item{EXT_METH}{DESCRIPTION}
\item{DAT_SET_CR}{DESCRIPTION}
\item{SRC_CITA}{DESCRIPTION}
\item{FIPS_ALPHA}{DESCRIPTION}
\item{NOAA_Regio}{DESCRIPTION}
\item{Shape_Leng}{DESCRIPTION}
\item{geometry}{DESCRIPTION}
}
}
}
\source{
\url{https://nsde.ngs.noaa.gov}
}
\usage{
nyc_shoclip_sf
}
\description{
Shoreline with parts of NJ and areas of Nassau County NY. sf object.
}
\details{
Manually adapted from NYC with part of Jersey shoreline and eastern coastline outside of NYC boundaries. Edited from CUSP sheet.
}
\author{
Kieran Healy
}
\keyword{datasets}
