% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nyc_state_senate_districts_sf}
\alias{nyc_state_senate_districts_sf}
\title{New York State Senate Districts}
\format{
\subsection{\code{nyc_state_senate_districts_sf}}{

A simple feature collection with 28 rows and 4 columns:
\describe{
\item{st_sen_dist}{State Senate district number}
\item{shape_leng}{Shape length}
\item{shape_area}{Shape area}
\item{geometry}{Multipolygon}
}
}
}
\source{
\url{https://www.nyc.gov/content/planning/pages/resources#datasets}
}
\usage{
nyc_state_senate_districts_sf
}
\description{
Simple feature collection of New York State Senate Districts within NYC.
EPSG:2263, NAD83 / New York Long Island (ftUS).
}
\details{
Boundaries of New York State Senate Districts within NYC. Produced by NYC
Planning Department. Release: 25C, August 2025.
}
\author{
Kieran Healy
}
\keyword{datasets}
