% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nyc_zcta_modzcta_df}
\alias{nyc_zcta_modzcta_df}
\title{NYC ZCTA to Modified ZCTA (MODZCTA) crosswalk}
\format{
\subsection{\code{nyc_zcta_modzcta_df}}{

A data frame with 215 rows and 2 columns:
\describe{
\item{zcta}{Census ZCTA (character)}
\item{modzcta}{Modified ZCTA the ZCTA belongs to (character). Joins to
\code{modzcta} in \link{nyc_modzcta_sf}.}
}
}
}
\source{
\url{https://github.com/nychealth/coronavirus-data/tree/master/Geography-resources}
}
\usage{
nyc_zcta_modzcta_df
}
\description{
Crosswalk from Census Zip Code Tabulation Areas (ZCTAs) to the NYC
Department of Health and Mental Hygiene's Modified ZCTAs.
}
\details{
Each row is one ZCTA. Several ZCTAs can map to the same MODZCTA. Use this
table to aggregate ZCTA-level data, such as Census or ACS estimates, to
MODZCTAs. Produced by NYC DOHMH.
}
\author{
Kieran Healy
}
\keyword{datasets}
