% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nyc_boros}
\alias{nyc_boros}
\title{New York City borough names}
\format{
\subsection{\code{nyc_boros}}{

A data frame with 5 rows and 5 columns:
\describe{
\item{boro_code}{City borough code. (Numeric)}
\item{boro_name}{Borough name}
\item{county_name}{County_Name County}
\item{short_county_name}{County_Name}
\item{long_county_name}{County_Name County, New York}
}
}
}
\source{
<SOURCE_URL>
}
\usage{
nyc_boros
}
\description{
Borough and county names and code.
}
\details{
DETAILS
}
\author{
Kieran Healy
}
\keyword{datasets}
