% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nyc_boros_sf}
\alias{nyc_boros_sf}
\title{New York City borough boundaries}
\format{
\subsection{\code{nyc_boros_sf}}{

A data frame with 5 rows and 5 columns:
\describe{
\item{boro_code}{Numeric borough code}
\item{boro_name}{Borough Name}
\item{shape_leng}{Shape length}
\item{shape_area}{Shape area}
\item{geometry}{Multipolygon}
}
}
}
\source{
\url{https://www.nyc.gov/content/planning/pages/resources#datasets}
}
\usage{
nyc_boros_sf
}
\description{
Clipped to shoreline. EPSG:2263, NAD83 / New York Long Island (ftUS). An sf object.
}
\details{
Borough boundaries for NYC, clipped to shoreline. Produced by NYC Planning Department. Release: 25C, August 2025.
}
\author{
Kieran Healy
}
\keyword{datasets}
